Fix race condition with default connectors, during reactor creation.
The `initSchedule` method can only prepare execution, but can never
actually cause execution to start. Only `schedule` can do that. To
ensure this, we had to change a bit how `Timer` scheduler works.

Once this is ensured, we only had to move the creation of default
connectors *after* the `initSchedule` call, and *immediately before*
`schedule`.

This makes `system.channels.await` work correctly.

Added a ScalaCheck property to check this.