Fix bug that allowed reactors to prematurely wake up a reactor.
During the initialization of the main and system channel in the
reactor system itself, the channel is now being shared. This happens
before the reactor gets scheduled, and potentially before the
corresponding `defaultConnector` and `internalConnector` fields
get set.

We now prevent other reactors from waking up the reactor -- the
reactor system (i.e. caller of `spawn`) is the only entity that
can trigger the first scheduling of the reactor.
This is achieved by setting the reactor's `active` field
before the channel references are shared.